steady ios/ipad versions
