+++ fullmoon/Views/Chat/ChatsListView.swift
@@ -1,5 +1,5 @@
 //
-//  ChatsView.swift
+//  ChatsListView.swift
 //  fullmoon
 //
 //  Created by Jordan Singer on 10/5/24.
@@ -8,7 +8,7 @@
 import SwiftUI
 import SwiftData
 
-struct ChatsView: View {
+struct ChatsListView: View {
     @EnvironmentObject var appManager: AppManager
     @Environment(\.dismiss) var dismiss
     @Binding var currentThread: Thread?
@@ -16,6 +16,7 @@ struct ChatsView: View {
     @Environment(\.modelContext) var modelContext
     @Query(sort: \Thread.timestamp, order: .reverse) var threads: [Thread]
     @State var search = ""
+    @State var selection: UUID?
     
     var body: some View {
         NavigationStack {
@@ -25,48 +26,50 @@ struct ChatsView: View {
                         Label(threads.count == 0 ? "no chats yet" : "no results", systemImage: "message")
                     }
                 } else {
-                    List {
-                        Section {
-                            ForEach(filteredThreads) { thread in
-                                Button {
-                                    setCurrentThread(thread)
-                                } label: {
-                                    VStack(alignment: .leading) {
-                                        Group {
-                                            if let firstMessage = thread.sortedMessages.first {
-                                                Text(firstMessage.content)
-                                                    .lineLimit(1)
-                                            } else {
-                                                Text("untitled")
-                                            }
-                                        }
-                                        .foregroundStyle(.primary)
-                                        .font(.headline)
-                                        
-                                        Text("\(thread.timestamp.formatted())")
-                                            .foregroundStyle(.secondary)
-                                            .font(.subheadline)
+                    List(selection: $selection) {
+                        ForEach(filteredThreads) { thread in
+                            VStack(alignment: .leading) {
+                                Group {
+                                    if let firstMessage = thread.sortedMessages.first {
+                                        Text(firstMessage.content)
+                                            .lineLimit(1)
+                                    } else {
+                                        Text("untitled")
                                     }
                                 }
-                                .tint(.primary)
+                                .foregroundStyle(.primary)
+                                .font(.headline)
+                                
+                                Text("\(thread.timestamp.formatted())")
+                                    .foregroundStyle(.secondary)
+                                    .font(.subheadline)
                             }
-                            .onDelete(perform: deleteThreads)
+                            .frame(maxWidth: .infinity, alignment: .leading)
+                            .contentShape(Rectangle())
+                            .onTapGesture {
+                                setCurrentThread(thread)
+                            }
+                            .tag(thread.id)
                         }
+                        .onDelete(perform: deleteThreads)
                     }
+                    .listStyle(.insetGrouped)
                 }
             }
             .searchable(text: $search, prompt: "search")
             .navigationTitle("chats")
             .navigationBarTitleDisplayMode(.inline)
             .toolbar {
-                ToolbarItem(placement: .topBarLeading) {
-                    Button(action: { dismiss() }) {
-                        Image(systemName: "xmark")
+                if appManager.userInterfaceIdiom == .phone {
+                    ToolbarItem(placement: .topBarLeading) {
+                        Button(action: { dismiss() }) {
+                            Image(systemName: "xmark")
+                        }
                     }
                 }
                 
                 ToolbarItem(placement: .topBarTrailing) {
-                    Button(action: { setCurrentThread() }) {
+                    Button(action: { setCurrentThread(nil) }) {
                         Image(systemName: "plus")
                     }
                 }
@@ -100,6 +103,11 @@ struct ChatsView: View {
     
     private func setCurrentThread(_ thread: Thread? = nil) {
         currentThread = thread
+        if let thread {
+            selection = thread.id
+        } else {
+            selection = nil
+        }
         isPromptFocused = true
         dismiss()
         appManager.playHaptic()
@@ -108,5 +116,5 @@ struct ChatsView: View {
 
 #Preview {
     @FocusState var isPromptFocused: Bool
-    ChatsView(currentThread: .constant(nil), isPromptFocused: $isPromptFocused)
+    ChatsListView(currentThread: .constant(nil), isPromptFocused: $isPromptFocused)
 }
